Move wait_for_crossbar fixture to util
